Update CHANGELOG
With this commit the clean up of Ground Zero aka rogue is finished. It
took me 10 month and about 150 hours. Every line of code was audited,
hundered of sanity checks added and of course all known bugs fixed.
Additionally the whole code was reformated. All in all Ground Zero
should now be much more reliable and or the first time in ~16 years
completeable without any crashes, logic bugs or the like.

To say some kind words: Ground Zero was by far the best game module code
that I've seen so far in my whole Quake II work. Especially the new
stuff like the Stalker or the Black Window was in much better shape that
anything else. Good work, whoever wrote it. :)

Of course this code is totaly untested. The next step is testing,
testing and testing.